requirements.txt
mysql-connector-python==8.0.31
mysqlclient==2.1.1
numpy==1.24.1
requests==2.28.1
flask==2.2.2
Flask-Session==0.4.0
Flask-MonitoringDashboard==3.1.1
opencv-python==4.6.0.66
mediapipe==0.9.0.1
tensorflow==2.11.0
multiprocess==0.70.14
matplotlib==3.6.2
unittest2==1.1.0
camera==1.3.0
opencv-camera==0.11.0
